backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=3.1.1
pymongo>=4.3.3
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
strawberry-graphql>=0.96.0
python-jose[cryptography]
passlib[bcrypt]
google-auth>=2.0.0
google-auth-oauthlib>=0.4.6
PyJWT==2.8.0